fix(components): mr review
